Moved the updateUnreadCount function to the ActionButtons component
to eliminate prop drilling and improve code organization maybe.
